% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIC_reconstruct.R, R/CIC_reconstruct_old.R
\name{CIC_reconstruct}
\alias{CIC_reconstruct}
\title{Using reconstruct individual-level data from KM_reconstruct and
digitized inputs from cumulative incidence curves, assign a cause
for each failure in competing risks situations. Only consider competing
risks, not semi-competing risks. One or two cumulative incidence curves.
One arm at a time.}
\usage{
CIC_reconstruct(IPD, clicks1, clicks2 = NULL)

CIC_reconstruct(IPD, clicks1, clicks2 = NULL)
}
\arguments{
\item{dat}{A data frame with arm, time, event.}

\item{aug_CIC}{A data frame processed through format_CIC().}

\item{dat}{A data frame with arm, time, event.}

\item{aug_CIC}{A data frame processed through format_CIC().}
}
\value{
An augmented version of dat that additionally gives the cause
of the event (cause 1 or cause 2).

An augmented version of dat that additionally gives the cause
of the event (cause 1 or cause 2).
}
\description{
Using reconstruct individual-level data from KM_reconstruct and
digitized inputs from cumulative incidence curves, assign a cause
for each failure in competing risks situations. Only consider competing
risks, not semi-competing risks. One or two cumulative incidence curves.
One arm at a time.

Using reconstruct individual-level data from KM_reconstruct and
digitized inputs from cumulative incidence curves, assign a cause
for each failure in competing risks situations. Only consider competing
risks, not semi-competing risks. One or two cumulative incidence curves.
One arm at a time.
}
\examples{
data(TTfields_pfs_trt_NAR)
data(TTfields_pfs_trt_clicks)
augmented_NAR <- format_NAR_tab(rawNAR=TTfields_pfs_trt_NAR, rawSurv=TTfields_pfs_trt_clicks)
KM_reconstruct(aug_NAR=augmented_NAR$aug_NAR, aug_surv=augmented_NAR$aug_surv)

data(TTfields_pfs_trt_NAR)
data(TTfields_pfs_trt_clicks)
augmented_NAR <- format_NAR_tab(rawNAR=TTfields_pfs_trt_NAR, rawSurv=TTfields_pfs_trt_clicks)
KM_reconstruct(aug_NAR=augmented_NAR$aug_NAR, aug_surv=augmented_NAR$aug_surv)

}
