% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparam_rmst.R
\name{integrate_survdat}
\alias{integrate_survdat}
\title{Calculate nonparametric RMST for data.frame with time, status, arm}
\usage{
integrate_survdat(dat, tau)
}
\arguments{
\item{dat}{Data frame of time-to-event data which MUST have the columns
'time', 'arm', and 'status}
}
\value{
list with rmst_pbo (arm=0) and rmst_trt (arm=1)
}
\description{
Calculate nonparametric RMST for data.frame with time, status, arm
}
\examples{
time <- rnorm(100)
status <- rbinom(n=100, size=0.5)
arm <- c( rep(1, 50), rep(0, 50))
dat <- data.frame(time=time, status=status, arm=arm)
integrate_survdat(dat=dat)

}
