% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM_reconstruct.R
\name{KM_reconstruct}
\alias{KM_reconstruct}
\title{Reconstruct individual-level data from augmented survival table and
NAR table, with augmentation performed by format_raw_tabs().}
\usage{
KM_reconstruct(aug_NAR, aug_surv)
}
\arguments{
\item{aug_NAR}{A data frame processed through format_raw_tabs().}

\item{aug_surv}{A data frame processed through format_raw_tabs().}
}
\value{
A list including IPD_time, IPD_event, n_hat=n_hat,
KM_hat, n_cen, n_event, int_censor
}
\description{
Reconstruct individual-level data from augmented survival table and
NAR table, with augmentation performed by format_raw_tabs().
}
\examples{
data(TTfields_pfs_trt_NAR)
data(TTfields_pfs_trt_clicks)
augmented_NAR <- format_NAR_tab(rawNAR=TTfields_pfs_trt_NAR, rawSurv=TTfields_pfs_trt_clicks)
KM_reconstruct(aug_NAR=augmented_NAR$aug_NAR, aug_surv=augmented_NAR$aug_surv)

}

