% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIC_reconstruct.R, R/CIC_reconstruct_old.R
\name{match_clicks}
\alias{match_clicks}
\title{Match clicks from cumulative incidence curve to clicks from KM curve.
We need the times from both sets of clicks to match exactly to guess
the number of events due to each cause.}
\usage{
match_clicks(km_dat, clicks_dat, prev_row = c(0, 0))

match_clicks(km_dat, clicks_dat, prev_row = c(0, 0))
}
\arguments{
\item{km_dat}{A data frame with time (and usually surv but that's not used here)}

\item{clicks_dat}{A data frame with time and est_cum_inc}

\item{prev_row}{The last row of the previous split (c(0,0) if not splitting and matching)}

\item{prev_cuminc}{The maximum est_cum_inc of the previous split (0 if not splitting and matching)}

\item{clicks_dat}{A data frame with time and est_cum_inc}

\item{km_dat}{A data frame with time (and usually surv but that's not used here)}
}
\value{
new_temp_clicks, the time-matched cum incidence clicks

new_temp_clicks, the time-matched cum incidence clicks
}
\description{
Match clicks from cumulative incidence curve to clicks from KM curve.
We need the times from both sets of clicks to match exactly to guess
the number of events due to each cause.

Match clicks from cumulative incidence curve to clicks from KM curve.
We need the times from both sets of clicks to match exactly to guess
the number of events due to each cause.
}
\examples{
data(TTfields_pfs_trt_NAR)
data(TTfields_pfs_trt_clicks)
augmented_NAR <- format_NAR_tab(rawNAR=TTfields_pfs_trt_NAR, rawSurv=TTfields_pfs_trt_clicks)
KM_reconstruct(aug_NAR=augmented_NAR$aug_NAR, aug_surv=augmented_NAR$aug_surv)
data(TTfields_pfs_trt_NAR)
data(TTfields_pfs_trt_clicks)
augmented_NAR <- format_NAR_tab(rawNAR=TTfields_pfs_trt_NAR, rawSurv=TTfields_pfs_trt_clicks)
KM_reconstruct(aug_NAR=augmented_NAR$aug_NAR, aug_surv=augmented_NAR$aug_surv)
}
