% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbo_clicks}
\alias{pbo_clicks}
\title{Pembrolizumab example OS KM reconstruction clicks - placebo arm}
\format{A data frame with 96 rows and 2 variables:
\describe{
  \item{time}{event time, months}
  \item{survival}{probability of progression-free survival}
}}
\source{
\url{NEJM. 2018;378(22):2078-2092}
}
\usage{
pbo_clicks
}
\description{
A dataset containing the clicks used to reconstruct the placebo OS KM curve.
}
\keyword{datasets}
