% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pembro_NAR}
\alias{pembro_NAR}
\title{Pembrolizumab example OS NAR table - pembrolizumab arm}
\format{A data frame with 8 rows and 2 variables:
\describe{
  \item{time}{event time, months}
  \item{NAR}{number of subjects at risk}
  ...
}}
\source{
\url{NEJM. 2018;378(22):2078-2092}
}
\usage{
pembro_NAR
}
\description{
A dataset containing the number at risk information for the pembrolizumab OS KM curve.
}
\keyword{datasets}
