% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PFS_noTTF_clicks}
\alias{PFS_noTTF_clicks}
\title{Tumor-Treating Fields example PFS KM reconstruction clicks - Temozolomide alone arm}
\format{
A data frame with 114 rows and 2 variables:
\describe{
  \item{time}{event time, months}
  \item{survival}{probability of progression-free survival}
}
}
\source{
\url{JAMA. 2018;318(23):2306-2316}
}
\usage{
PFS_noTTF_clicks
}
\description{
A dataset containing the clicks used to reconstruct the Temozolomide alone KM curve.
}
\keyword{datasets}
