% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PFS_noTTF_NAR}
\alias{PFS_noTTF_NAR}
\title{Tumor-Treating Fields example PFS NAR table - Temozolomide alone arm}
\format{
A data frame with 6 rows and 2 variables:
\describe{
  \item{time}{event time, months}
  \item{NAR}{number of subjects at risk}
  ...
}
}
\source{
\url{JAMA. 2018;318(23):2306-2316}
}
\usage{
PFS_noTTF_NAR
}
\description{
A dataset containing the number at risk information for the Temozolomide alone KM curve.
}
\keyword{datasets}
