% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PFS_TTF_NAR}
\alias{PFS_TTF_NAR}
\title{Tumor-Treating Fields example PFS NAR table - TTFields + temozolomide arm}
\format{
A data frame with 6 rows and 2 variables:
\describe{
  \item{time}{event time, months}
  \item{NAR}{number of subjects at risk}
  ...
}
}
\source{
\url{JAMA. 2018;318(23):2306-2316}
}
\usage{
PFS_TTF_NAR
}
\description{
A dataset containing the number at risk information for the TTFields + temozolomide KM curve.
}
\keyword{datasets}
