% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_cox_outputs.R
\name{print_cox_outputs}
\alias{print_cox_outputs}
\title{Just a wrapper to get quantities out of a call to coxph()}
\usage{
print_cox_outputs(cox_fit, print_output = TRUE)
}
\arguments{
\item{cox_fit}{A model fitted with coxph()}
}
\value{
A list including IPD_time, IPD_event, n_hat=n_hat,
KM_hat, n_cen, n_event, int_censor
}
\description{
Just a wrapper to get quantities out of a call to coxph()
}
\examples{
time <- rnorm(100)
status <- rbinom(n=100, size=0.5)
arm <- c(rep(1,50), rep(0,50))
temp_cox <- coxph(Surv(time, status) ~ arm)
print_cox_outputs(temp_cox)

}

