% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TTfields_pfs_pbo_clicks}
\alias{TTfields_pfs_pbo_clicks}
\title{Tumor-Treating Fields PFS KM reconstruction clicks - placebo}
\format{A data frame with 114 rows and 2 variables:
\describe{
  \item{time}{event time, months}
  \item{survival}{probability of progression-free survival}
}}
\source{
\url{JAMA. 2018;318(23):2306-2316}
}
\usage{
TTfields_pfs_pbo_clicks
}
\description{
A dataset containing the clicks used to reconstruct the placebo KMC in Fig 2A
}
\keyword{datasets}
