% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TTfields_pfs_trt_NAR}
\alias{TTfields_pfs_trt_NAR}
\title{Tumor-Treating Fields PFS NAR table - treatment}
\format{A data frame with 6 rows and 2 variables:
\describe{
  \item{time}{event time, months}
  \item{NAR}{number of subjects at risk}
  ...
}}
\source{
\url{JAMA. 2018;318(23):2306-2316}
}
\usage{
TTfields_pfs_trt_NAR
}
\description{
A dataset containing the number at risk information for treatment KMC in Fig 2A
}
\keyword{datasets}
