% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM_reconstruct.R
\name{KM_reconstruct}
\alias{KM_reconstruct}
\title{KM_reconstruct.R}
\usage{
KM_reconstruct(aug_NAR, aug_surv)
}
\arguments{
\item{aug_NAR}{A data frame processed through format_raw_tabs().}

\item{aug_surv}{A data frame processed through format_raw_tabs().}
}
\value{
A list including IPD_time, IPD_event, n_hat=n_hat,
KM_hat, n_cen, n_event, int_censor
}
\description{
Reconstruct individual-level data from augmented survival table and
augmented NAR table, with augmentation performed by format_raw_tabs().
}
\examples{
data(PFS_TTF_NAR)
data(PFS_TTF_clicks)
augTabs <- format_raw_tabs(raw_NAR=PFS_TTF_NAR, raw_surv=PFS_TTF_clicks)
KM_reconstruct(aug_NAR=augTabs$aug_NAR, aug_surv=augTabs$aug_surv)

}
