% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pembro_clicks}
\alias{pembro_clicks}
\title{Pembrolizumab example OS KM reconstruction clicks - pembrolizumab arm}
\format{
A data frame with 97 rows and 2 variables, time (event time in months) and survival (probability of OS)
}
\usage{
data(pembro_clicks)
}
\description{
A dataset containing the clicks used to reconstruct the pembrolizumab OS KM curve.
}
\references{
Gandhi et al. NEJM 2018;378(22):2078-2092
}
\keyword{datasets}
